refactor: MakeThreadFormを分解し状態に関する部分をMakeThreadへ
