+++ src/containers/MakeThread.js
@@ -1,14 +1,122 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
 
-import MakeThreadForm from '../components/MakeThreadForm';
+import {
+  TextFields,
+  ImageButton,
+  Checkbox,
+  OtherButtons
+} from '../components/MakeThreadForm';
 import * as channelActions from '../modules/channelModule';
 
 function MakeThread(props) {
   const { addThread } = props;
+  // TextFieldsに渡す
+  const [title, setTitle] = useState('');
+  const [details, setDetails] = useState('');
+  const [isTitleFilled, setIsTitileFilled] = useState(true); // 訪問した最初にはエラーは出さない
+
+  // ImageButtonに渡す
+  const [pictureURL, setPictureURL] = useState('');
+
+  // checkBoxに渡す
+  const [isFirst, setIsFirst] = useState(false);
+  const [isSecond, setIsSecond] = useState(false);
+  const [isThird, setIsThird] = useState(false);
+  const [isFourth, setIsFourth] = useState(false);
+  const [isFifth, setIsFifth] = useState(false);
+  const [isFastHalf, setIsFastHalf] = useState(false);
+  const [isFastEnd, setIsFastEnd] = useState(false);
+  const [isLateHalf, setIsLateHalf] = useState(false);
+  const [isLateEnd, setIsLateEnd] = useState(false);
+
+  function handleTextChange(event) {
+    const targetName = event.target.name;
+    const value = event.target.value;
+    switch (targetName) {
+      case 'title':
+        setTitle(value);
+        // 何か入力されていればerror表示を消す
+        if (value.trim() !== '') {
+          setIsTitileFilled(true);
+        }
+        return;
+      case 'details':
+        setDetails(value);
+        return;
+      default:
+        return;
+    }
+  }
+
+  function handleCheckChange(event) {
+    const targetName = event.target.name;
+    const checked = event.target.checked;
+    switch (targetName) {
+      case 'first':
+        setIsFirst(checked);
+        return;
+      case 'second':
+        setIsSecond(checked);
+        return;
+      case 'third':
+        setIsThird(checked);
+        return;
+      case 'fourth':
+        setIsFourth(checked);
+        return;
+      case 'fifth':
+        setIsFifth(checked);
+        return;
+      case 'fastHalf':
+        setIsFastHalf(checked);
+        return;
+      case 'fastEnd':
+        setIsFastEnd(checked);
+        return;
+      case 'lateHalf':
+        setIsLateHalf(checked);
+        return;
+      case 'lateEnd':
+        setIsLateEnd(checked);
+        return;
+      default:
+        return;
+    }
+  }
+
+  function handleSubmit() {
+    // タイトルの入力欄が空だったりホワイトスペースばっかりだったら送信しない
+    // String.tirm() で文字列の銭湯と最後にある改行は空白を取り除
+    if (title.trim() === '') {
+      setIsTitileFilled(false);
+      return;
+    }
+    addThread(title.trim(), details.trim(), pictureURL);
+  }
+
   return (
     <React.Fragment>
-      <MakeThreadForm addThread={addThread} />
+      <TextFields
+        title={title}
+        details={details}
+        isTitleFilled={isTitleFilled}
+        onChange={handleTextChange}
+      />
+      <ImageButton />
+      <Checkbox onChange={handleCheckChange} />
+      <OtherButtons
+        isFirst={isFirst}
+        isSecond={isSecond}
+        isThird={isThird}
+        isFourth={isFourth}
+        isFifth={isFifth}
+        isFastHalf={isFastHalf}
+        isFastEnd={isFastEnd}
+        isLateHalf={isLateHalf}
+        isLateEnd={isLateEnd}
+        onSubmit={handleSubmit}
+      />
     </React.Fragment>
   );
 }
